requirements.txt
pyspark==3.5.1
delta-spark==3.1.0
pandas
requests
beautifulsoup4
psycopg2-binary